PR #1145: Source changes for review

+++ src/psyclone/tests/test_files/dynamo0p3/15.28.1_int_sign_X_builtin.f90
@@ -43,7 +43,7 @@ program single_invoke
   implicit none
 
   type(integer_field_type) :: f1, f2
-  real(i_def)              :: a
+  integer(i_def)           :: a
 
   a = 1
